kino: version bumped to 1.3.3.
Changes:
* New DVD Video authoring script that launches the Brasero burning application
* New export script for the PLAYSTATION 3
* New Chinese (Traditional) translation
* Updated Catalan translation

NOTE: The default value for the audio device is changed to /dev/dsp (OSS) to be
more compatible with PulseAudio, which is suddenly very commonplace. It is
compatible with most ALSA setups through the OSS emulation modules. It bypasses
PulseAudio by default, whose ALSA module does not work well with Kino. Finally,
in this configuration, it is easy to make it use PulseAudio with A/V synch
using the padsp wrapper. If you do not use PulseAudio and just want to use ALSA
then you can still set the Audio Device to "default" or some other PCM name
in Kino's Preferences.