tellico: version bumped to 1.3.5.
Entry updating was fixed so that it does not combine multiple values.
The CueCat decoder was updated to work for both UPC and ISBN searches.
The Delicious Library importer was updated to import games and movies.
A query bug with z39.50 ISBN searches was fixed.
Using images in a local data directory no longer marks every collection as modified.
A crash when exporting to a small Alexandria collection was fixed.