stellarium: version bumped to 0.10.2.
This version is based on the recent release of Qt 4.5, which should fix
startup problems which some users experienced with version 0.10.1.
This change should also solve some Mac rendering bugs and speed up the
GUI elements of the program. A new projection mode is now available,
called "Hammer-Aitoff", which allows for very wide fields of view (up to
360 degrees). This release also includes many bugfixes and some general
performance tweaks.