lyx: version bumped to 1.6.2.
This is the second maintenance release in the 1.6.x series. The release
fixes a large number of major and critical bugs that were reported by
users of LyX 1.6.0 and 1.6.1. Most notably, three bugs that might lead to
dataloss were detected and resolved. Furthermore, a number of crashes,
file conversion and LaTeX output problems are fixed. As a bonus, the
performance of the application has been improved, and the release also
introduces some new features and user interface improvements.