asterisk: a new module. There are a few things missing, like
some ISDN stuff, Radius, Voicetronix and a couple of others. I could't
figure a sane way to do the /etc/asterisk conf files and settled on
this approach. If anyone has a better idea, feel free to implement.

This is module isn't really tested as I have no free box to really
give it a go with. But it does compile fine on x86_64, so I presume
x86 will be fine as well.

Don't bother with --with-blah. It leads down a nasty road of you
needing to specify paths and just compilcates things.