mkvtoolnix: version bumped to 2.7.0.
Several bugs with QuickTime audio and video codecs, incorrect timestamps
and an inactive progress bar in mmg's job manager, and problems with
quoting and escaping of mkvmerge's arguments have been fixed. The header
editor supports dragging and dropping of files. The header editor writes
changes in a way that is more compatible with existing parsers.
Support for TrueHD/MLP audio and the "forced track" flag was added.
The aspect ratio extraction for AVC/h.264 video was fixed, and aspect ratio
information is now kept in the bitstream by default.
mkvmerge uses SimpleBlock elements by default.