syslinux: version bumped to 3.75.
This is a collection of fixes over the 3.74 release. The most important
ones are fixing the geometry detection in MEMDISK and fixing the PXELINUX
"keeppxe" option, which was broken in 3.74. The MENU SAVE behavior has
been improved, gPXELINUX no longer disables interrupts when it shouldn't,
and a lockup bug in the Hardware Detection Tool has been addressed.