qt4: version bump.
Needed to alter the build since they added an extra question you need to
answer to proceed with configure, that's the purpose of the COMMERCIAL_USER
set.

Additionally, for future versions of Amarok4, you'll need the qtscriptgenerator,
which needs the -exceptions. Providing a choice for those who don't want the
extra bloat. You won't need to say "y" to the docs, examples, etc.