bluez: version bump.
The Bluetooth SIG announced the 3.0 + HS specification that includes support for
High Speed data transfer via 802.11 technology. Using a Bluetooth 2.1 BR/EDR
controller with an 802.11 AMP controller is a perfect combination for a 3.0 high
speed solution. So BlueZ is fully ready to be extended with 3.0 support since it
already supports Bluetooth 2.1 features.