e_dbus-svn: Re-arrange DEPENDS.
In current moonbase, e17-svn does pull in HAL optionally. This does not
have any effect in the first place, as e_dbus is the module linking
against HAL if the user wants it to.

I also added a CONFIGURE for some build options available which do not
seem to serve any purpose yet, but in future versions these network
options might get a backend for enlightenemnt. The commented lines in
CONFIGURE should be taken care of if something like that happens.

Note: downloads and compiles OK, and e17-svn builds on top of it
      but also installs /usr/share/doc/e_dbus-svn/doc/images/.svn!
      untested by me - engelsman