sqlite: version bumped to 3.6.14.
Changes:
Added the optional asynchronous VFS module.
Enhanced the query optimizer so that virtual tables are able to make
use of OR and IN operators in the WHERE clause.
Speed improvements in the btree and pager layers.
Added the SQLITE_HAVE_ISNAN compile-time option which will cause the
isnan() function from the standard math library to be used instead of
SQLite's own home-brew NaN checker.
Countless minor bug fixes, documentation improvements, new and improved
test cases, and code simplifications and cleanups.