rox-filer: version bumped to 2.8.
Fix the install. The MIME icon naming was updated to the latest freedesktop.org
scheme. This should make GNOME themes work better with ROX-Filer. A work-around
was added for a panel placement problem with Compiz-fusion 0.7.4. The "Path"
key in .desktop files is now supported. The Vietnamese, Russian, Italian, and
French translations were updated, and some bugs were fixed.