rosegarden: updated to version 1.73
From the web page:
"This is primarily a bug fix/maintenance release. It addresses some
critical track-related crashes that have been affecting users over the
last couple of releases, and introduces assorted other minor fixes,
such as improved multi-track recording for transposing instruments. It
also includes the incomplete beginnings of several improvement projects
that have all been postponed for now while we focus all our attention
on the port to Qt4.

We hope for this to be the final release of the KDE/Qt3-based
Rosegarden. All work on this line of code has been suspended, and all
of our effort is going into the new Qt4-based Rosegarden "Thorn" which
we hope to release at some point in 2009. Unless we discover some truly
critical problems with this release, any bugs and usability complaints
will be deferred until after Thorn has entered the beta testing phase,
and all new work will be done on the new code. "