opencv: version bump and a number of changes to BUILD/DEPENDS.
octave and OpenEXR (ImfCRgbaFile) causes the make to barf, so holding
off on those for now. As noted for ffmpeg vs gstreamer, its one or the
other. Odd that xine-lib is not included with that but there you have it.

opencv devs are now aware of the pathing issues with ffmpeg and libdc1394.