graphviz: version bump, a couple of more optional_depends, and a few other
tweaks. The patch takes care of gv.pm and gv.so getting symlinked in /root because
PERL_INSTALL_DIR was not getting set in the tclpkg/Makefile.

The additions to BUILD fixes the relin that happens because libperl.so is not found.
You will get the same kick off if you have R installed as well.