poppler: version bumped to 0.11.1.
Many features were added, such as support for colorizing text in pattern
colorspace, the possibility of forcing no hinting of fonts in the Splash
backend, support for multiple page sizes when converting to PostScript,
and handling of Streams in CMap definitions.
More work was done on Annotations support.
Several minor bugs were fixed.
The GLib and Qt4 frontends were improved, along with their demos and documentation.