ekiga: version bumped to 3.2.5.
This release fixed a very low rate of sent images during video conversations,
a crash in LDAP, a crash in presence, a crash during idle times, a crash when
accepting a call, a deadlock in the SIP handler, and a freeze upon calling
and showing of pc:udp$... in the URI bar.
Multiple registrations with the same registrar using different user names is
now allowed.
Compilation with GTK+ 2.12 and GCC 4.4 was fixed.
Translations were updated.