sudo:version bumped to 1.7.2.
This release introduces a new #includedir directive to sudoers, which
can be used to implement an /etc/sudo.d directory.
User and group names in sudoers may now be enclosed in double quotes
to avoid having to escape special characters.
Bugfixes include a fix for the -g option when only setting the group,
a hang in visudo checking aliases, support for shell-style
/etc/environment files on Linux, BSM audit fixes when changing to a
non-root UID, setenv() compatibility fixes for Linux systems, and a
work-around for certificate and key file problems on Netscape-derived
LDAP SDKs.