gambas2: version bumped to 2.15.0.
A big bug was fixed in the interpreter that made it crash randomly on some systems.
String to number conversion was fixed, especially when converting Long numbers.
The Firebird driver is compiled correctly now.
The SQLite 3 driver now correctly returns the default value of fields.
Raising or lowering controls now works correctly in the GTK+ component.
The SMTP component now strictly follows RFC 2047 when sending mails with subjects encoded in UTF-8.
The Xml.ToString() method was fixed in the XML component.