ghostscript: version bumped to 8.70.
This release fixes a large number of issues with transparency, especially
as it interacts with color space conversion, mask contexts and patterns.
There have also been a number of significant fixes to font handling,
especially when generating PDF. And there are numerous robustness,
correctness and performance improvements.
New generic Esc/Page devices, eplmono and eplcolor were added to the
contrib directory. A new cdnj500 device was added to support the HP
DesignJet 500.
The licensing of the Free version of the core Ghostscript code has been
changed to GPLv3 or later. Previously, the core code was GPLv2 only.
This release also includes security fixes addressing CVE-2009-0583 and
CVE-2009-0792.