hdparm: version bumped to 9.22.
Support for SATA SSD TRIM commands was enhanced with the introduction
of several arguments. In addition, an experimental script (wiper.sh)
is provided which calculates a list of unallocated blocks within a
filesystem, and informs the SSD firmware of those blocks. In some cases,
this can restore a sluggish SSD to nearly-new speeds again.
Furthermore, a workaround to an observed ext4 bug with FIEMAP's "LAST"
flag was added, and the "-E" flag was modified to work with modern DVD
drives.