poppler: version bumped to 0.11.3.
Several optimizations in the Cairo renderer.
A fix in alpha handling of some files in the Splash rendering.
Conversion to PS in some files has been fixed.
Better locale handling.
The addition of the Color Management API to the Qt4 frontend.
The -png flag has been added to pdftoppm to output to PNG.