libisoburn: version bumped to 0.4.2.
The xorriso program in this release can record MD5 checksums for the
whole session and for each single data file. One may verify data file
content in the image. When comparing image files with files in the
local filesystem, the MD5 sums avoid the need for reading file content
from the image.