Subject: bzedit2: different source url
download failed during an 'lget -a' sweep.  old source url is dead.
version unchanged, so no need to change UPDATED.

also under the impression that bzedit no longer required for creating
bzflag landscapes because it can be done in blender, but I could be wrong.
If bzedit is removed from moonbase, then gtkglarea could probably go too.

Note: downloads, compiles and installs OK,
      but untested as I don't have GLX extension installed - engelsman

PS. discovered that there is an out-of-date [non-lunar] lunar source
archive, which is cool, so could just point to that.