Subject: BitTorrent: updated to version 5.2.0
the previous version failed to download during an 'lget -a' sweep.
this still isn't the most up-to-date version: there is a binary
download for 6.2c for Windows, and various pre-built packages, but
no obvious source code.

as with the previous version, the bittorrent-ncurses runs but the
GUI version doesn't find the wx import, even though wxPython is
now a dependency!