ethumb-svn: new module.
I really have no clue at the moment how to utilize it correctly and
it seems to be pretty new (alpha-ish). I added it mainly for e_modules-svn
to compile correctly. Its configure script does not report it missing
if it's not installed, but the build will fail without out.

There's also an optional dependency, called epdf-svn, which would offer
a PDF viewer and PDF thumbnail generation by ethumb. The current epdf-svn
does require poppler, but doesn't seem to like the one we have in moonbase,
so I suggest to keep it on hold for now until the E folks resume updating it.