phpMyAdmin: version bumped to 3.2.2.
When sorting, the red arrows were reversed in the list of tables: this
was fixed. One could not edit a row without a primary key and a BIT
field: this was fixed as well. The results of a query which contains a
LIMIT clause inside a subquery could not be exported and this has been
fixed.