ekiga: version bumped to 3.2.6.
This release fixes various crashes and freezes.
It adds a workaround for registrars that refuse registration when several contacts are given.
It fixes "Contacts never go offline".
It makes sure the output device is set before a call.
It fixes port handling when registering to a proxy with a non-standard port.
It is possible to register e.g. user@a.b as a username as required by some providers.
There are many other fixes.