libexif: version bumped to 0.6.18.
This release's highlights include greatly expanded API documentation
and sample programs, improved support for Pentax, Casio, and Epson
MakerNotes, increased stability in the face of corrupted JPEG files,
and proper output alignment in UTF-8 locales. EXIF tag fixup is now
more eager in the default case, which means that more mandatory tags
are added when needed and others are automatically corrected to be of
the proper data types. New translations for the da, is, it, lv, pt_BR,
and sr locales are also included.