sqlite: version bumped to 3.6.19.
Changes:
Added support for foreign key constraints.
Foreign key constraints are disabled by default.
Use the foreign_keys pragma to turn them on.
Generalized the IS and IS NOT operators to take arbitrary expressions
on their right-hand side.
The TCL Interface has been enhanced to use the Non-Recursive Engine
(NRE) interface to the TCL interpreter when linked against TCL 8.6 or
later.
Fix a bug introduced in 3.6.18 that can lead to a segfault when
an attempt is made to write on a read-only database.