poppler: version bumped to 0.12.1
Changes:
Compilation with some compilers was fixed.
A crash on fonts without CharCodeToUnicode was resolved.
Reliability for streams with broken Length was improved.
A crash when saving files that come from a stream without a name was resolved.
The reliability of the save function was improved.
An incorrect Length value is now fixed when saving.
Security fixes based on xpdf 3.02pl4 were integrated.
Several other bugfixes and improvements were made.
The build system was enhanced.