gambas2: version bumped to 2.18.0.
Changes:
The IDE has support for the Ukrainian language.
It does not sometimes freeze anymore when running a project.
Native windows are now disabled with GTK+ 2.18 so that the GTK+ component
works correctly on recent distributions (Ubuntu 9.10, Mandriva 2010.0, and so on).
Control.Reparent() now works correctly in the GTK+ component.
The RpcServer class of the XML/RPC component now correctly handles callbacks
having more than one argument.