sendmail: version bumped to 8.14.4.
This version fixed a problem where some certificate authorities do not
properly check the requests they are signing and hence allow spoofing
via an embedded NUL in the CN entry. A workaround for a Linux resolver
problem has been added to avoid core dumps. A bug where the value of
headers, e.g., Precedence, Content-Type, et.al., was not extracted
correctly thus preventing them from being recognized properly was
fixed. An erroneous reduction of the length limitation on a return path
was fixed.