eric4: version bumped to 4.4.0.
Changes:
A Python3 compatible debugger and a respective project type were added.
Debugger support was improved with a a thread list viewer for the debug
viewer, support for forking to the Python debuggers, and code for
handling infinite recursions to the Python debuggers.
The help viewer was changed to a full blown Web browser (based on QWebkit).
The SQL Browser tool was added to inspect databases.
A little icon editor tool was added.
Additional translations were added.
Qt3 support was removed from eric4.
A lot of other changes were made.