atop: version bumped to 1.24.
The screen interface has been redesigned completely with support for a
variable number of columns depending on the window size. It is now
possible to define your own output line with process-level statistics.
A system-wide configuration file was introduced for atop and atopsar.
Handling of log files was improved and additional counters are maintained.