fsarchiver: version bumped to 0.6.6.
This release will not remove the archive file when savefs/savedir fails
because the archive already exists. Partitions already mounted are
remounted with MS_BIND to have access to files hidden by mounted
filesystems. This release will analyze filesystems only when they are
all accessible to prevent having to wait and then getting an error.
There are many internal changes, and an improved layout on the main screen.