fsarchiver: version bumped to 0.6.7.
Support for sparse files was added, so a sparse file on the original
disk will be recreated as sparse file. Progression information was
implemented for savedir/restdir. The logfile created when the -d
option is used now has a unique name so that it's not overwritten.
The MD5 checksums are now calculated using the implementation from
libgcrypt instead of the internal one, so libgcrypt is always
required. Other internal changes were made.