virtuoso: Version bump. Please note the lower part of the DETAILS for additional
information. This version is not compatable with any prior versions. Since kde
uses this mainly for PIM stuff, so if you have any indexed data you might want to
back it up some where... just in case.

I was going to add a virtuoso converter to moonbase that is supposed to make this
conversion process easier but have read reports that is did not work for most people.
Also the converter wanted to download virtuoso-1.5.x, why I don't know but thought it
to be overkill if you already had 1.5.x installed... so I nixed that idea.