glibmm: updated to version 2.22.1
second attempt at this update, because last time I somehow
managed to say I was updating to 2.22.1, but in fact I was
downgrading from 2.22.1 to 2.20.1, and then got a bit lost
when a subsequent merge with 2.20.0 failed.

Moral of the tale: don't do clever things with git after
midnight when you're tired and should already be in bed!