netpbm: Version bump. Adding a CONFIGURE and adjusting the BUILD to allow
for compiling without X11; this was noted by Richard Pyne having a machine
where X11 was not an option.

It seems to work here. If "n" is said, then there is no;

X11LIB = /usr/X11R6/lib/libX11.so

in the Makefile.config. Saying "y" and that variable is present.