upstart: an alternative to init.
Continuing my theme of replacing low-level base functionality with
new and adventurous things, here's upstart, Ubuntu's replacement for
init.  It's largely a drop-in replacement--for those times that it
isn't, I've included base configuration files so that you'll still
be able to log in.

It installs to its default /usr/sbin/init instead of the system-default
/sbin/init so that it doesn't conflict with sysvinit, giving people a
chance to try both of them out without having to remove one.  If this
becomes a proper replacement for init, I expect that its install
location will change to the normal /sbin.