mlt: Version bump with the following;
"This is a minor maintenance release, but it is interesting because it now
enables usage of libswscale as the default choice for image scaling, image
format conversion, and color space conversion. That gives better quality and
performance. In addition, there are some improvements in the sdl_preview
consumer to make it suitable for use in OpenShot 1.1."