smb4k4: version bumped to 0.10.6.
Changes:
Include files are no longer installed.
The Portuguese translation was updated.
A problem with shares that have no comment was fixed.
Listing of shares when a "creating lame ..." message appears on stderr was fixed.
smb4k_umount was backported from SVN HEAD.
Under non-FreeBSD operating systems, it falls back to umount if umount.cifs can not be found.
The check for mount and umount binaries was removed from Smb4KCore::searchPrograms().