nano: version bumped to 2.2.4.
This release includes security fixes due to an assessment of nano's
vulnerability to symlink attacks; see CVE-2010-1160 and CVE-2010-1161.
Also included are fixes for various crashes when using the spell
checker on new files, and fixing the 'file was modified' message when
saving to a new filename. 'Final' fixes to page up/down due to the soft
wrapping code are included as well. The lone new feature included is a
new syntax highlighting definition for cmake files.