Revert "lcms: updated to version 2.0"
This reverts commit abf70549f7ad532fc6af0379007f165a70aa89d1.

No module that uses this is capable of detecting lcms presense because its headers,
libs and lcms.pc have a "2". After some attempts at renaming for example lcms2.h to
lcms.h and others in a similar fashion. To much has changed and the three or so apps such
as, libmng, poppler, and ImageMagick will fail their makes on internal changes to the
headers and libs.

This module should not be bumped at this time until the 15 or so existing modules are
ready for it.