hdparm: version bumped to 9.29.
The "--offset" flag was added for use with "-t", to get timings from
the middle or end of a drive. The "--trim-sector-ranges-stdin" was
restricted to small TRIM commands on most drives. Verbose reporting
was modified. Reporting of SATA3 signaling speed was added to "-I".
The use of ATA_16 for all commands to ATAPI devices is now forced.
A sector count of "1" is now supplied to IDENTIFY commands, to
satisfy buggy USB3 bridges.