Subject: ftgl: version update
The package still hasn't been updated in two years, but this update
is better than the one from 2006. The only package that depends on
ftgl is libprojectM and the version of it in the moonbase is no
longer supported on Sourceforge at the moment. It can safely be
upgraded separately.

I ran into issues compiling the lightspark Flash plugin, which
depended on this slightly newer version of the library.

Note: downloads, compiles, installs OK, but untested here.
      No NVIDIA on this box, so that part untested too - engelsman