less: removed LESSCHARSET variable defining LESSCHARSET to latin1 is very bad and makes no sense. Text displayed via less is terribly broken if not using old ISO8859 encoding or US ASCII. Also makes no sense. Here is info why: From the manpage:
If neither LESSCHARSET nor LESSCHARDEF is set, but any of the strings
"UTF-8", "UTF8", "utf-8" or "utf8" is found in the LC_ALL, LC_TYPE or
LANG environment variables, then the default character set is utf-8.

If that string is not found, but your system supports the setlocale
interface, less will use setlocale to determine the character set.
setlocale is controlled by setting the LANG or LC_CTYPE environment
variables.

Finally, if the setlocale interface is also not available, the default
character set is latin1.

So removing LESSCHARSET from profile fixes display on my utf-8 Lunar.