aria2: version bumped to 1.01.0.
This release adds an option to limit the number of connections to the
same host in each download. aria2 now chooses the server which is least
used in the aria2c instance. This release also adds Chromium cookie
support and HTTP only conditional download support in which a file is
downloaded only when the local file is older than the remote file. aria2
now can handle %2F in FTP URI properly.
The HTTP/1.1 chunked decoder was fixed.