sudo: version bumped to 1.7.4p2.
A bug where sudo could spin in a busy loop waiting for the child process
was fixed. A bug introduced in sudo 1.7.3 that prevented the -k and -K
options from functioning when the tty_tickets sudoers option is enabled
was fixed. Sudo no longer prints a warning when the -k or -K options are
specified and the ticket file does not exist.