vlc: version bumped to 1.1.3.
Some more bugs were squashed, and a potential security issue leading
to a crash was discovered in the taglib module. This release
introduces small fixes and improvements over VLC 1.1.2 for most
platforms, notably for FTP, DVD, and podcast. It provides updated
translations for Dutch, Sinhala, Hebrew, Estonian, Spanish, and
Bengali.