gtksourceview-sharp: tweaked BUILD to disable sample dir
added sedit to BUILD to disable build of sample subdirectory,
otherwise it fails with errors about being unable to find
PrintJob and PrintDialog functions. Searching the web shows
various people with the same problem [in spanish!] but no
clear solution, or gives so many hits about usage [in Java!]
as to be useless.

Who uses this mono stuff anyway?