xine-lib-vdpau: This started out as a simple lin, except the autogen.sh/configure
would crap out with a missing Makefile.in. After several manual downloads of the
source, kept getting the same error. So I did a svn co from their repository and
recreated the tarball which corrected that missing file.

The patch from xine-lib needs to be applied here to fix a build error from the
bumped videoproto. Along the way it appears to be PSAFE (here anyway) and the
commented CFLAGS can be removed from the BUILD.