x11vnc: version bumped to 0.9.12.
A compile time error when using --with-system-libvncserver to point to
LibVNCServer 0.9.7 was fixed. A bug from the forced use of Xdefs.h is
worked around. The kdm display manager is now detected when trying not
to get killed by the display manager. One can now specify the maximum
number of displays that can be created in -create mode via
X11VNC_CREATE_MAX_DISPLAYS.