qt4: Version bump. There are a number of changes here in the BUILD. I am fairly certain these path
changes allow for the coexistance between the qt provided phonon and the phonon moudle. As such it
will effect the phonon module (relin it). If you do an update or manually, the easiest thing to do
is relin automoc which should take care of much of the kde modules. As always lunar fix is your friend.