qt4: After Ratler was so kind to fix a ld_remove issue that I bellyached about
I could stop being a knucklehead and actually use it where needed. As mentioned
by zbiggy in the mail list Qt will try and use an existing install which causes
things to tank. So lets twiddle with the LDFLAGS path before the BUILD starts.

Also, wekbit is not needed as an optional_depends, the option is to have qt
build its own webkit plugin. Let there be rejoicing.