xwininfo: version bumped to 1.1.0.
The major new feature of this release is the rewrite to use libxcb instead
of libX11, batching requests for information to reduce waiting on round-trips
in gathering the client information from the X server.   Testing over a
high-latency connection from California to China and back resulted in an
improvement from 8 minutes to 45 seconds to output the complete window tree
via xwininfo -root -all of a GNOME desktop session with 114 windows.