feh: version bumped to 1.10.
It is now possible to display arbitrary information in the image by using
the --info option. Minor workarounds and improvements for scrolling and
Xinerama handling have been made. This is also the first release in which
the graphical (X11) parts of feh have been tested in a number of ways.