PyQt4: version bumped to 4.8.
Qt 4.7 is now supported.
pyqtProperty() can now be used as a method decorator in the same way
that the standard Python property type can be used. Unbound signals
can now be subscripted in the same way that bound signals can.
The missing << operator was added to QPolygon, QPolygonF, and
QXmlStreamAttributes. PyQt will now automatically keep a reference to
objects set with some sort of setter method when Qt expects the
underlying C++ instance to stay valid but does not take ownership of
it. lightmaps.py and stardelegate.py examples were added. The
--assume-shared option was added to configure.py