aliases: Adding an alias "%JPEG" and adjusting the appropriate DEPENDS.
Like past jpeg version bumps, this is not exactly a clean process if you
switch to libjpeg-turbo; it of course depends on what you have installed.
So like in the past jpeg bumps, some manual intervention maybe required.

The vast majority of this list I do have installed and found no issues
during the lin or useage.

Or you can just stay with jpeg and avoid the whole thing.