xf86dga: version bumped to 1.0.3.
This release raises the dependency on libXxf86dga to require a minimum
of version 1.1.0 to get the new Xxf86dga.h header instead of the now
deprecated xf86dga.h.  However, it also removes unused dependencies on
libXaw, libXt, and libXmu that were previously declared.
It also delivers the build configuration improvements and other
janitorial cleanups.