poppler: DEPENDS update and new CONFIGURE
changed lcms and cairo to opt deps and added missing opt deps

the recommendations in DEPENDS are directly from popplers configure script

I couldn't figure out, whether the options I included in CONFIGURE
have dependencies. if so, please move them to the DEPENDS script.

I toyed a little with the options to maybe figure out how to build
epdfview, but haven't succeeded. maybe someone else finds out and adds
a comment to the right options or so, would be nice.