gnome-commander: version bumped to 1.2.8.10.
This release fixes a problem with numeric keypad arrows, a problem with
metatag popup menu, a problem with toggling path/basename/filename
selections in copy/move dialogs, and a problem with searching path for
devices. A menu item was added for copying full path and file name to the
clipboard. It also updates the German translation.