netpbm: build update
I fixed the compilation errors I encountered during build.
those were due to the interactive configure script:
apparently they changed if you type none at an opt dep there will be
no need for a default after it.

I also removed CONFIGURE. Since you can take the options out of
DEPENDS. I used OPTS for that to create the interactive options IACT.
I revised the whole creation of the IACT string, very verbose so it'll
be easier to change it in the future.

I also left libpng and added jbigkit as well as jasper, which are all
in Makefile.config.in but not in the interactive configure script,
but I haven't written any sedit Makefile changes.