poppler: version bumped to 1.16.1.
A colorspace issue in the Cairo backend was fixed.
A crash on failure of jpeg_create_decompress(), as well as a crash in
filling certain forms were fixed. An overlay bug in the generation of
Level 1 separable PostScript was resolved.
A leak in the Splash backend and a leak in the Glib frontend for pages
with no text were fixed.
Other minor fixes were made.