tcp_wrappers: Switching the SOURCE2 patch with the one from the LFS folks.
http://www.linuxfromscratch.org/patches/blfs/svn/tcp_wrappers-7.6-shared_lib_plus_plus-1.patch

This started off with a module I plan to add; dcmtk, which tanked on the wrappers. Their
INSTALL says it will only work with tcp_wrappers that are ANSI C compliant. Seems
our patches lack in that area.

A `lunar fix` found no issues and I have recompiled xinetd, inetutils, openldap and openssh with
no problems. At Ratlers request, the hosts.deny has been verified to work.

Bumping the UPDATED because of the patch change and what it does.