coreutils: version bumped to 8.10.
There are some minor bug fixes, along with two new features. The join
feature is enabled via a new option, "-o auto". The cp feature makes
copying sparse files much more efficient on several common file systems.
It takes advantage of a feature that was introduced in linux-2.6.27.
The improvement affects the default code path, so if you're looking
for risk potential, this is it.  It uses the feature if available,
and otherwise resorts to using the old, less-efficient copying code.