qt4: Version bump. Some folks have had build issues with this and it was
suggested to use a PRE_BUILD as done with heimdal. Not been able to duplicate
that problem but adding a PRE_BUILD anyway. Hope it fixes their problems.

Also, the prior qt4+phonon and kde still exists. Hopefully that will be fixed with
the next KDE4 bump or phonon bump.