texlive: using only system libs/utils now ...
texlive uses either system or internal libs/utils, so if one does not
have/enable an optional dependency texlive would build its own version
of it, which is kinda not what we want.

This patch explicitly sets all dependencies to required, that would be
built by texlive and explicitly turns the internal building of them of,
see the huge OPTS+= in BUILD.

The only libs missing, i.e. not in moonbase and built by texlive, are:
- kpathsea
- ptexenc