kde4/base: This is part of the fleshing out of DEPENDS. I hadn't planned on doing
this right now but kdebindings pushed me into it. Was trying to cypher why kdebindings
was bombing out on qscintilla2. Which by the way is not at all unusual with the
Riverbank trio; sip, PyQt4 and/or qscintilla2. So for now say no to qscintilla2 and qwt.

Didn't want to lrm qscintilla2 so added the pair value to the DEPENDS but it would still
barf but adding it to the BUILD worked. Long story short; the tripping point is how we
call cmake, the OPTS and do the out of source build. With the later being the real issue.

For now kdelibs4 is the only one that needs/wants an out of source build and all the rest
do not. So the BUILDS have been adjusted for that with an additional change to the cmake
build lines. The cmake man states it should be called in this way;

cmake [options] <path-to-source>

while we have the two reversed. So I have changed it such that we call cmake with its OPTS
first and then the source path. The easiest way to determine if something wants an out of
source build; untar it, cd to its top level, run `ccmake .` then press `c` and it will let
you know.

It probably doesn't matter much but well, there you have it.