kdesdk4: Changing the build method to out of source build. This one is an exception
to the other kde4/base modules. As mentioned in commit c03b0717ff94794d1fb7f1835aec371f70ff4c05
to determine which method is best; untar, cd to dir, ccmake ., press c and it will
say it wants an out of source build. This module did not give that indication and was failing
the build on some kate/plugins.