emacs: install emacs's "ctags" program as "etags"
This allows you to have both the "ctags" and "emacs" package, for people
who like using all of the tools from both.  "etags" is a traditional
name for emacs's "ctags" command, so it won't be too unexpected for
experienced emacs users (I hope).  If it turns out to be problematic, it
can always be made a build-time option.