kde4: Version bump to 4.7.0. A lot of changes here. First off I am using the same naming
conventions as shown in the kde repository; I see no reason to complicate that aspect by
retaining things like kdebase4-runtime and then go "Oh yeah, that's really kde-runtime." pfft. So
this begins the process of renaming or rather dropping the "4". Right now I have decided to keep all
these in base for the sole reason that is how they are in their repository; that may change at a
later date.

I am not real crazy about having all these listed in the DEPENDS of the kde profile but it is
what it is.

Also this begins the dropping of the $MAJOR.

The mono stuff I have not tested, nor will I so for those folks that have problems with it, work it
out and commit.

Unlike the audio issue mentioned by zbiggy, it did not happen here. It *might* be that I have switched
this box to pure systemd, I dunno. OTOH he is right about that ConsoleKit stuff; it can slow down your login
and cause some delay in opening applications. I have tried the consolekit_add_with_parameters.patch but it
did not solve anything. Not a big deal, kde will work without it.