qtzeitgeist: This is essentially a rename of libqzeitgeist. They moved the repository from
git://gitorious.org/kde-zeitgeist/libqzeitgeist.git to https://projects.kde.org/projects/kdesupport/libqzeitgeist/repository.
The problem; gitorious was deleted and libqzeitgest (their current master) causes the phonon make to
barf because the libqzeitgeist includes are a bit hosed.

Afaik they will be renaming to qtzeitgeist anyway in the near future so in the mean time we will
use the debian orig version which works with our current phonon.