gcc: The short story; ntl barfed looking for /usr/lib/gcc/x86_64-pc-linux-gnu/4.5.2/crtbeginS.o
even though I had gcc-4.5.3 installed. A quick look in /usr/lib/gcc/x86_64-pc-linux-gnu and
there were remnants of old gcc's laying around each with a symlink pointing to
/usr/bin/cpp. In fact the only thing in there was a cpp.

Putting the contents of the POST_INSTALL in each of the BUILDs and removing the PSOT_INSTALL.

Think very carefully of what you put in POST_INSTALL because as in this case; file changes
will not be tracked.