pinentry: Adjusting the DEPENDS. Making ncurses a hard depends to be used as a
fallback. The --enable-pinentry-qt switch is for qt3 so don't try poking that
in the optional_depends entry, it will tank the ./confiugre. There is no qt4
switch.

Adjusting the BUILD to make ncurses the fallback and adding a if/then should
qt4 be enabled in the DEPENDS to set the profile rc.